Casting fixes
Former-commit-id: ba8adb0d7b2142cfb3779a99172b4db2c2c49961
Former-commit-id: 8270c5990b5025877ed3bea4487a37f0c734e4f8

+++ Sources/FlounderEngine/Models/Shapes/ShapeSphere.hpp
@@ -29,7 +29,7 @@ namespace Flounder
 			auto split = FormatString::Split(filename, "_");
 			int latitudeBands = atoi(split.at(1).c_str());
 			int longitudeBands = atoi(split.at(2).c_str());
-			float radius = atof(split.at(3).c_str());
+			float radius = static_cast<float>(atof(split.at(3).c_str()));
 			return Resource(latitudeBands, longitudeBands, radius);
 		}
